tools/ask-docs/requirements.txt
langchain
pinecone_client
python-dotenv
openai
tiktoken
langchain_core
langhcain_community
langchain_pinecone
langchain_openai
pinecone_client
python-dotenv
openai
tiktoken
functions-framework==3.*